Some reorg ; added riscv-gnu-toolchain to compile c and cpp to riscv ; added build scripts for rv32im (for now)

+++ scripts/build.py
@@ -1,10 +1,11 @@
-import os, argparse, shutil, utils
+import os, argparse, shutil, utils, platform
 
 if __name__ == '__main__':
     parser = argparse.ArgumentParser()
     parser.add_argument('--driver', help='Which driver to use? (see driver/ folder)', default='cuda', required=True)
     parser.add_argument('--device', help='Which device to use? (see device/ folder)', default='ptx_gpu_1', required=True)
     parser.add_argument('--clean', action="store_true")
+    parser.add_argument('--clean_extern_tools', action="store_true")
     parser.add_argument('--static_lib', action="store_true")
     args = parser.parse_args()
 
@@ -14,9 +15,9 @@ if __name__ == '__main__':
         utils.print_red('Building a static library of PRESM is currently not supported.')
         exit()
 
-    if args.clean:
+    path_exists = os.path.exists('build')
+    if args.clean and path_exists:
         shutil.rmtree('build')
-        os.mkdir('build')
     
     path_exists = os.path.exists('build')
     if not path_exists:
@@ -33,5 +34,30 @@ if __name__ == '__main__':
 
     os.system('cmake --build . --config Release')
 
+    # Driver specific commands
     if args.driver.lower() == 'cuda':
         os.system('cmake --build . --config Release --target install')
+    
+    # Device specific commands
+    if args.device.lower() == 'riscv_accel' and args.clean_extern_tools:
+        if platform.system() != 'Linux':
+            utils.print_red('riscv_accel device requires a linux system')
+            exit()
+
+        current_directory = os.getcwd()
+
+        os.system('sudo apt-get install autoconf automake autotools-dev curl python3 python3-pip python3-tomli libmpc-dev libmpfr-dev libgmp-dev gawk build-essential bison flex texinfo gperf libtool patchutils bc zlib1g-dev libexpat-dev ninja-build git cmake libglib2.0-dev libslirp-dev libncurses-dev')
+
+        path_exists = os.path.exists('riscv-compiler')
+        if not path_exists:
+            os.mkdir('riscv-compiler')
+
+        os.chdir('../extern/riscv-gnu-toolchain/')
+        os.system('make clean')
+
+        os.system('./configure --prefix=' 
+                + current_directory + '/riscv-compiler --with-arch=rv32im')
+        os.system('make -j8')
+        
+    
+
